fixed the tripwire command to include the string init, improved file structure, changed yargs call
